@@ -1,25 +1,19 @@
 name: viam_example_app
-description: An example app using Viam's Flutter SDK
-publish_to: "none" # Remove this line if you wish to publish to pub.dev
-
-version: 1.0.0+1
+description: A new Flutter project.
+publish_to: "none"
+version: 0.1.0
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.1.3 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  flutter_dotenv: ^5.1.0
   viam_sdk:
     path: ../../
-  cupertino_icons: ^1.0.2
-  flutter_platform_widgets: ^3.2.1
-  flutter_webrtc: ^0.9.35
-  image: ^4.0.17
-  flutter_dotenv: ^5.1.0
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.13.1
   flutter_test:
     sdk: flutter
   flutter_lints: ^2.0.0
@@ -28,21 +22,3 @@ flutter:
   uses-material-design: true
   assets:
     - .env
-
-flutter_launcher_icons:
-  android: "launcher_icon"
-  ios: true
-  image_path: "assets/icon/app_logo.png"
-  min_sdk_android: 21 # android min sdk min:16, default 21
-  web:
-    generate: true
-    image_path: "assets/icon/app_logo.png"
-    background_color: "#hexcode"
-    theme_color: "#hexcode"
-  windows:
-    generate: true
-    image_path: "assets/icon/app_logo.png"
-    icon_size: 48 # min:48, max:256, default: 48
-  macos:
-    generate: true
-    image_path: "assets/icon/app_logo.png"
